pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Solovtcov Stanislav <vandemar@bk.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.1.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
flask = "^2.3.3"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.7"
validators = "^0.22.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
html5lib = "^1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
